Address comments on code review.
