+++ idiss-csharp/IdissLib/Types.cs
@@ -1,9 +1,6 @@
-﻿using System;
-using System.Runtime.InteropServices;
-using System.Text;
-using System.Text.Json;
+
+using System;
 using System.Collections.Generic;
-using System.Text.Json.Serialization;
 
 namespace IdissLib
 {
@@ -342,193 +339,8 @@ namespace IdissLib
     public class IpPrivateKeys
     {
         /// The private key used to sign the identity object sent back to the user.
-        public string ip_private_key { get; set; }
+        public string ipPrivateKey { get; set; }
         /// The private key used to sign the initial account creation message.
-        public string ip_cdi_private_key { get; set; }
-    }
-
-
-    /// A JsonConverter for JSON serialization and deserialization of objects of type Dictionary<AttributeTag, Attribute>
-    // It is public so that in case be used in tests and from applications using the library.
-    public class DictionaryConverter : JsonConverter<Dictionary<AttributeTag, Attribute>>
-    {
-        public override Dictionary<AttributeTag, Attribute> Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
-        {
-            if (reader.TokenType != JsonTokenType.StartObject)
-            {
-                throw new JsonException();
-            }
-            var value = new Dictionary<AttributeTag, Attribute>();
-            while (reader.Read())
-            {
-                if (reader.TokenType == JsonTokenType.EndObject)
-                {
-                    return value;
-                }
-                string keyString = reader.GetString();
-                var key = new AttributeTag(keyString);
-                reader.Read();
-                string itemString = reader.GetString();
-                var itemValue = new Attribute(itemString);
-                value.Add(key, itemValue);
-            }
-            throw new JsonException("Error Occured");
-        }
-
-        public override void Write(Utf8JsonWriter writer, Dictionary<AttributeTag, Attribute> value, JsonSerializerOptions options)
-        {
-            writer.WriteStartObject();
-            foreach (KeyValuePair<AttributeTag, Attribute> item in value)
-            {
-                writer.WriteString(item.Key.tag, item.Value.attribute);
-            }
-            writer.WriteEndObject();
-        }
-    }
-
-    /// A JsonConverter for JSON serialization and deserialization of objects of type YearMonth.
-    public class YearMonthConverter : JsonConverter<YearMonth>
-    {
-        public override YearMonth Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
-        {
-            string str = reader.GetString();
-            return new YearMonth(str);
-        }
-
-        public override void Write(Utf8JsonWriter writer, YearMonth value, JsonSerializerOptions options)
-        {
-            writer.WriteStringValue(value.str);
-        }
-    }
-
-
-    /// A JsonConverter for JSON serialization and deserialization of objects of type AccountAddress.
-    public class AccountAddressConverter : JsonConverter<AccountAddress>
-    {
-        public override AccountAddress Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
-        {
-            string address = reader.GetString();
-            return new AccountAddress(address);
-        }
-
-        public override void Write(Utf8JsonWriter writer, AccountAddress value, JsonSerializerOptions options)
-        {
-            writer.WriteStringValue(value.address);
-        }
-    }
-
-    /// An Exception to be thrown in case of validation failure of a request.
-    public class RequestValidationException : Exception
-    {
-        public RequestValidationException(string message) : base(message)
-        {
-        }
-    }
-
-    /// An Exception to be thrown in case that identity creation does not succeed.
-    public class IdentityCreationException : Exception
-    {
-        public IdentityCreationException(string message) : base(message)
-        {
-        }
-    }
-
-    public static class Idiss
-    {
-        /// Import of the two C functions that are exported by the Rust library "idiss". 
-        [DllImport("idiss.dll")]
-        private static extern IntPtr validate_request_cs([MarshalAs(UnmanagedType.LPArray)] byte[] ctx, int ctx_len,
-        [MarshalAs(UnmanagedType.LPArray)] byte[] ip_info, int ip_info_len,
-        [MarshalAs(UnmanagedType.LPArray)] byte[] ars_infos, int ars_infos_len,
-        [MarshalAs(UnmanagedType.LPArray)] byte[] request, int request_len, out int out_length, out int out_success);
-
-        [DllImport("idiss.dll")]
-        private static extern IntPtr create_identity_object_cs([MarshalAs(UnmanagedType.LPArray)] byte[] ip_info, int ip_info_len,
-        [MarshalAs(UnmanagedType.LPArray)] byte[] request, int request_len,
-        [MarshalAs(UnmanagedType.LPArray)] byte[] alist, int alist_len,
-        UInt64 expiry,
-        [MarshalAs(UnmanagedType.LPArray)] byte[] ip_private_key, int ip_private_key_ptr_len,
-        [MarshalAs(UnmanagedType.LPArray)] byte[] ip_cdi_private_key, int ip_cdi_private_key_ptr_len,
-        out int out_length, out int out_success);
-
-        /// A wrapper around the imported C function "validate_request_cs". The arguments are
-        /// - global - the versioned global context
-        /// - ip_info - the versioned identity provider info
-        /// - ars_info - the versioned map of ArInfos
-        /// - request - the identity object request
-        /// The function either
-        /// - returns the address of the initial account, if the request is valid, or
-        /// - throws an exception, if the request is invalid or the input is malformed. 
-        public static AccountAddress validate_request(VersionedGlobalContext global, VersionedIpInfo ip_info, VersionedArInfos ars_infos, IdObjectRequest request)
-        {
-            byte[] global_bytes = JsonSerializer.SerializeToUtf8Bytes(global);
-            byte[] request_bytes = JsonSerializer.SerializeToUtf8Bytes(request);
-            byte[] ars_infos_bytes = JsonSerializer.SerializeToUtf8Bytes(ars_infos);
-            byte[] ip_info_bytes = JsonSerializer.SerializeToUtf8Bytes(ip_info);
-            int out_length = 0;
-            int out_success = 0;
-            var result_ptr = validate_request_cs(global_bytes, global_bytes.Length, ip_info_bytes, ip_info_bytes.Length, ars_infos_bytes, ars_infos_bytes.Length, request_bytes, request_bytes.Length, out out_length, out out_success);
-            byte[] out_bytes = new byte[out_length];
-            Marshal.Copy(result_ptr, out_bytes, 0, out_length);
-            if (out_success == 1)
-            {
-                return new AccountAddress(Encoding.UTF8.GetString(out_bytes));
-            }
-            else if (out_success == -1)
-            {
-                var error_string = Encoding.UTF8.GetString(out_bytes);
-                throw new RequestValidationException(error_string);
-            }
-            else
-            {
-                throw new RequestValidationException("Unknown error");
-            }
-        }
-
-        /// A wrapper around the imported C function "create_identity_object_cs". The arguments are
-        /// - ip_info - the versioned identity provider info
-        /// - alist - the attribute list containing the attributes of the user
-        /// - request - the identity object request. This is the request that the wallet sends which contains cryptographic values and proofs.
-        /// - expiry - the expiry time of the account creation message sent to the chain. This is just a unix timestamp that should be set to, e.g., now + 5 min.
-        /// - ip_keys - the private keys of the identity provider
-        /// The function either
-        /// - returns a `IdentityCreation` object containing 
-        ///     * the identity object that is returned to the user
-        ///     * the anonymity revocation record
-        ///     * the initial account creation object that is sent to the chain
-        ///     * the address of the inital account, or 
-        ///  - throws an exception, if any of the inputs are malformed. 
-
-        public static IdentityCreation create_identity_object(VersionedIpInfo ip_info, AttributeList alist, IdObjectRequest request, UInt64 expiry, IpPrivateKeys ip_keys)
-        {
-            var options = new JsonSerializerOptions();
-            options.Converters.Add(new DictionaryConverter());
-            options.Converters.Add(new YearMonthConverter());
-            options.Converters.Add(new AccountAddressConverter());
-            byte[] request_bytes = JsonSerializer.SerializeToUtf8Bytes(request);
-            byte[] ip_info_bytes = JsonSerializer.SerializeToUtf8Bytes(ip_info);
-            byte[] alist_bytes = JsonSerializer.SerializeToUtf8Bytes(alist, options);
-            byte[] ip_private_key_bytes = Encoding.UTF8.GetBytes(ip_keys.ip_private_key);
-            byte[] ip_cdi_private_key_bytes = Encoding.UTF8.GetBytes(ip_keys.ip_cdi_private_key);
-            int id_out_length = 0;
-            int out_success = 0;
-            var id_ptr = create_identity_object_cs(ip_info_bytes, ip_info_bytes.Length, request_bytes, request_bytes.Length, alist_bytes, alist_bytes.Length,
-             expiry, ip_private_key_bytes, ip_private_key_bytes.Length, ip_cdi_private_key_bytes, ip_cdi_private_key_bytes.Length, out id_out_length, out out_success);
-            byte[] id_out_bytes = new byte[id_out_length];
-            Marshal.Copy(id_ptr, id_out_bytes, 0, id_out_length);
-            if (out_success == 1)
-            {
-                return JsonSerializer.Deserialize<IdentityCreation>(id_out_bytes, options);
-            }
-            else if (out_success == -1)
-            {
-                var error_string = Encoding.UTF8.GetString(id_out_bytes);
-                throw new IdentityCreationException(error_string);
-            }
-            else
-            {
-                throw new IdentityCreationException("Unkown error.");
-            }
-        }
+        public string ipCdiPrivateKey { get; set; }
     }
 }
